Added user registration, authentication and some extra features.

+++ week2/model.php
@@ -19,18 +19,19 @@ error_reporting(E_ALL);
  * @param string $pass database password
  * @return pdo object
  */
-function connect_db($host, $db, $user, $pass){
+function connect_db($host, $db, $user, $pass)
+{
     $charset = 'utf8mb4';
 
     $dsn = "mysql:host=$host;dbname=$db;charset=$charset";
     $options = [
-        PDO::ATTR_ERRMODE            => PDO::ERRMODE_EXCEPTION,
+        PDO::ATTR_ERRMODE => PDO::ERRMODE_EXCEPTION,
         PDO::ATTR_DEFAULT_FETCH_MODE => PDO::FETCH_ASSOC,
     ];
     try {
         $pdo = new PDO($dsn, $user, $pass, $options);
     } catch (\PDOException $e) {
-        echo sprintf("Failed to connect. %s",$e->getMessage());
+        echo sprintf("Failed to connect. %s", $e->getMessage());
     }
     return $pdo;
 }
@@ -42,9 +43,10 @@ function connect_db($host, $db, $user, $pass){
  * @return bool
  *
  */
-function new_route($route_uri, $request_type){
+function new_route($route_uri, $request_type)
+{
     $route_uri_expl = array_filter(explode('/', $route_uri));
-    $current_path_expl = array_filter(explode('/',parse_url($_SERVER['REQUEST_URI'], PHP_URL_PATH)));
+    $current_path_expl = array_filter(explode('/', parse_url($_SERVER['REQUEST_URI'], PHP_URL_PATH)));
     if ($route_uri_expl == $current_path_expl && $_SERVER['REQUEST_METHOD'] == strtoupper($request_type)) {
         return True;
     }
@@ -56,7 +58,8 @@ function new_route($route_uri, $request_type){
  * @param bool $active Set the navigation item to active or inactive
  * @return array
  */
-function na($url, $active){
+function na($url, $active)
+{
     return [$url, $active];
 }
 
@@ -65,7 +68,8 @@ function na($url, $active){
  * @param string $template filename of the template without extension
  * @return string
  */
-function use_template($template){
+function use_template($template)
+{
     $template_doc = sprintf("views/%s.php", $template);
     return $template_doc;
 }
@@ -75,15 +79,16 @@ function use_template($template){
  * @param array $breadcrumbs Array with as Key the page name and as Value the corresponding url
  * @return string html code that represents the breadcrumbs
  */
-function get_breadcrumbs($breadcrumbs) {
+function get_breadcrumbs($breadcrumbs)
+{
     $breadcrumbs_exp = '
     <nav aria-label="breadcrumb">
     <ol class="breadcrumb">';
     foreach ($breadcrumbs as $name => $info) {
-        if ($info[1]){
-            $breadcrumbs_exp .= '<li class="breadcrumb-item active" aria-current="page">'.$name.'</li>';
-        }else{
-            $breadcrumbs_exp .= '<li class="breadcrumb-item"><a href="'.$info[0].'">'.$name.'</a></li>';
+        if ($info[1]) {
+            $breadcrumbs_exp .= '<li class="breadcrumb-item active" aria-current="page">' . $name . '</li>';
+        } else {
+            $breadcrumbs_exp .= '<li class="breadcrumb-item"><a href="' . $info[0] . '">' . $name . '</a></li>';
         }
     }
     $breadcrumbs_exp .= '
@@ -93,11 +98,12 @@ function get_breadcrumbs($breadcrumbs) {
 }
 
 /**
- * Creates navigation HTML code using given array
- * @param array $navigation Array with as Key the page name and as Value the corresponding url
+ * Creates navigation HTML code using given array with active_id page
+ * @param array $template Array with as Key the page name and as Value the corresponding url
+ * @param integer $active_id the id of the active element in the navigation bar
  * @return string html code that represents the navigation
  */
-function get_navigation($navigation){
+function get_navigation($template, $active_id){
     $navigation_exp = '
     <nav class="navbar navbar-expand-lg navbar-light bg-light">
     <a class="navbar-brand">Series Overview</a>
@@ -106,16 +112,16 @@ function get_navigation($navigation){
     </button>
     <div class="collapse navbar-collapse" id="navbarSupportedContent">
     <ul class="navbar-nav mr-auto">';
-    foreach ($navigation as $name => $info) {
-        if ($info[1]){
+    foreach ($template as $route_id => $array_info) {
+        if ($route_id == $active_id) {
             $navigation_exp .= '<li class="nav-item active">';
-            $navigation_exp .= '<a class="nav-link" href="'.$info[0].'">'.$name.'</a>';
-        }else{
-            $navigation_exp .= '<li class="nav-item">';
-            $navigation_exp .= '<a class="nav-link" href="'.$info[0].'">'.$name.'</a>';
+            $navigation_exp .= '<a class="nav-link" href="' . $array_info['url'] . '">' . $array_info['name'] . '</a>';
         }
+        else{
+            $navigation_exp .= '<li class="nav-item">';
+            $navigation_exp .= '<a class="nav-link" href="'.$array_info['url'].'">'.$array_info['name'].'</a>';
+    }
 
-        $navigation_exp .= '</li>';
     }
     $navigation_exp .= '
     </ul>
@@ -126,27 +132,33 @@ function get_navigation($navigation){
 
 /**
  * Creats a Bootstrap table with a list of series
- * @param object $db pdo object
- * @param array $series with series from the db
+ * @param array $series pdo object
+ * @param array $names_user with names from users from the db
  * @return string
  */
-function get_serie_table($series){
+function get_serie_table($series, $names_user)
+{
+    $i = 0;
     $table_exp = '
     <table class="table table-hover">
     <thead
     <tr>
         <th scope="col">Series</th>
+        <th scope="col">Added by</th>
         <th scope="col"></th>
     </tr>
     </thead>
     <tbody>';
-    foreach($series as $key => $value){
+    foreach ($series as $key => $value) {
+//        foreach ($names_user as $user => $names) {
         $table_exp .= '
         <tr>
-            <th scope="row">'.$value['name'].'</th>
-            <td><a href="/DDWT18/week2/serie/?serie_id='.$value['id'].'" role="button" class="btn btn-primary">More info</a></td>
+            <th scope="row">' . $value['name'] . '</th>
+            <td>' . $names_user[$i]['firstname'] . ' ' . $names_user[$i]['lastname'] . '</td>
+            <td><a href="/DDWT18/week2/serie/?serie_id=' . $value['id'] . '" role="button" class="btn btn-primary">More info</a></td>
         </tr>
         ';
+        $i += 1;
     }
     $table_exp .= '
     </tbody>
@@ -159,7 +171,8 @@ function get_serie_table($series){
  * Pritty Print Array
  * @param $input
  */
-function p_print($input){
+function p_print($input)
+{
     echo '<pre>';
     print_r($input);
     echo '</pre>';
@@ -170,14 +183,15 @@ function p_print($input){
  * @param object $pdo database object
  * @return array Associative array with all series
  */
-function get_series($pdo){
+function get_series($pdo)
+{
     $stmt = $pdo->prepare('SELECT * FROM series');
     $stmt->execute();
     $series = $stmt->fetchAll();
     $series_exp = Array();
 
     /* Create array with htmlspecialchars */
-    foreach ($series as $key => $value){
+    foreach ($series as $key => $value) {
         foreach ($value as $user_key => $user_input) {
             $series_exp[$key][$user_key] = htmlspecialchars($user_input);
         }
@@ -191,14 +205,15 @@ function get_series($pdo){
  * @param int $serie_id id from the serie
  * @return mixed
  */
-function get_serieinfo($pdo, $serie_id){
+function get_serieinfo($pdo, $serie_id)
+{
     $stmt = $pdo->prepare('SELECT * FROM series WHERE id = ?');
     $stmt->execute([$serie_id]);
     $serie_info = $stmt->fetch();
     $serie_info_exp = Array();
 
     /* Create array with htmlspecialchars */
-    foreach ($serie_info as $key => $value){
+    foreach ($serie_info as $key => $value) {
         $serie_info_exp[$key] = htmlspecialchars($value);
     }
     return $serie_info_exp;
@@ -210,10 +225,12 @@ function get_serieinfo($pdo, $serie_id){
  * @param string $message Error/Success message
  * @return string
  */
-function get_error($feedback){
+function get_error($feedback)
+{
+    $feedback = json_decode($feedback, True);
     $error_exp = '
-        <div class="alert alert-'.$feedback['type'].'" role="alert">
-            '.$feedback['message'].'
+        <div class="alert alert-' . $feedback['type'] . '" role="alert">
+            ' . $feedback['message'] . '
         </div>';
     return $error_exp;
 }
@@ -222,9 +239,11 @@ function get_error($feedback){
  * Add serie to the database
  * @param object $pdo db object
  * @param array $serie_info post array
+ * @param integer $user_id user that adds the series
  * @return array with message feedback
  */
-function add_serie($pdo, $serie_info){
+function add_serie($pdo, $serie_info, $user_id)
+{
     /* Check if all fields are set */
     if (
         empty($serie_info['Name']) or
@@ -250,7 +269,7 @@ function add_serie($pdo, $serie_info){
     $stmt = $pdo->prepare('SELECT * FROM series WHERE name = ?');
     $stmt->execute([$serie_info['Name']]);
     $serie = $stmt->rowCount();
-    if ($serie){
+    if ($serie) {
         return [
             'type' => 'danger',
             'message' => 'This series was already added.'
@@ -258,21 +277,21 @@ function add_serie($pdo, $serie_info){
     }
 
     /* Add Serie */
-    $stmt = $pdo->prepare("INSERT INTO series (name, creator, seasons, abstract) VALUES (?, ?, ?, ?)");
+    $stmt = $pdo->prepare("INSERT INTO series (name, creator, seasons, abstract, user) VALUES (?, ?, ?, ?, ?)");
     $stmt->execute([
         $serie_info['Name'],
         $serie_info['Creator'],
         $serie_info['Seasons'],
-        $serie_info['Abstract']
+        $serie_info['Abstract'],
+        $user_id
     ]);
     $inserted = $stmt->rowCount();
-    if ($inserted ==  1) {
+    if ($inserted == 1) {
         return [
             'type' => 'success',
             'message' => sprintf("Series '%s' added to Series Overview.", $serie_info['Name'])
         ];
-    }
-    else {
+    } else {
         return [
             'type' => 'danger',
             'message' => 'There was an error. The series was not added. Try it again.'
@@ -284,9 +303,18 @@ function add_serie($pdo, $serie_info){
  * Updates a serie in the database using post array
  * @param object $pdo db object
  * @param array $serie_info post array
+ * @param integer $user_id user id from the session info
  * @return array
  */
-function update_serie($pdo, $serie_info){
+function update_serie($pdo, $serie_info, $user_id)
+{
+    /* Check if the user is allowed to edit the serie */
+    if ($serie_info['user'] =! $user_id){
+        return [
+            'type' => 'danger',
+            'message' => 'There was an error. You are not allowed to edit this series.'];
+    }
+
     /* Check if all fields are set */
     if (
         empty($serie_info['Name']) or
@@ -319,7 +347,7 @@ function update_serie($pdo, $serie_info){
     $stmt = $pdo->prepare('SELECT * FROM series WHERE name = ?');
     $stmt->execute([$serie_info['Name']]);
     $serie = $stmt->fetch();
-    if ($serie_info['Name'] == $serie['name'] and $serie['name'] != $current_name){
+    if ($serie_info['Name'] == $serie['name'] and $serie['name'] != $current_name) {
         return [
             'type' => 'danger',
             'message' => sprintf("The name of the series cannot be changed. %s already exists.", $serie_info['Name'])
@@ -333,16 +361,16 @@ function update_serie($pdo, $serie_info){
         $serie_info['Creator'],
         $serie_info['Seasons'],
         $serie_info['Abstract'],
-        $serie_info['serie_id']
+        $serie_info['serie_id'],
     ]);
+
     $updated = $stmt->rowCount();
-    if ($updated ==  1) {
+    if ($updated == 1) {
         return [
             'type' => 'success',
             'message' => sprintf("Series '%s' was edited!", $serie_info['Name'])
         ];
-    }
-    else {
+    } else {
         return [
             'type' => 'warning',
             'message' => 'The series was not edited. No changes were detected'
@@ -356,21 +384,28 @@ function update_serie($pdo, $serie_info){
  * @param int $serie_id id of the to be deleted series
  * @return array
  */
-function remove_serie($pdo, $serie_id){
+function remove_serie($pdo, $serie_id, $user_id)
+{
     /* Get series info */
     $serie_info = get_serieinfo($pdo, $serie_id);
+
+    /* Check if the user is allowed to edit the serie */
+    if ($serie_info['user'] =! $user_id){
+        return [
+            'type' => 'danger',
+            'message' => 'There was an error. You are not allowed to edit this series.'];
+    }
 
     /* Delete Serie */
     $stmt = $pdo->prepare("DELETE FROM series WHERE id = ?");
     $stmt->execute([$serie_id]);
     $deleted = $stmt->rowCount();
-    if ($deleted ==  1) {
+    if ($deleted == 1) {
         return [
             'type' => 'success',
             'message' => sprintf("Series '%s' was removed!", $serie_info['name'])
         ];
-    }
-    else {
+    } else {
         return [
             'type' => 'warning',
             'message' => 'An error occurred. The series was not removed.'
@@ -383,7 +418,8 @@ function remove_serie($pdo, $serie_id){
  * @param object $pdo database object
  * @return mixed
  */
-function count_series($pdo){
+function count_series($pdo)
+{
     /* Get series */
     $stmt = $pdo->prepare('SELECT * FROM series');
     $stmt->execute();
@@ -392,10 +428,25 @@ function count_series($pdo){
 }
 
 /**
+ * Count the number of users that are registered on Series Overview
+ * @param object $pdo database object
+ * @return mixed
+ */
+function count_users($pdo)
+{
+    /* Get users */
+    $stmt = $pdo->prepare('SELECT * FROM users');
+    $stmt->execute();
+    $users = $stmt->rowCount();
+    return $users;
+}
+
+/**
  * Changes the HTTP Header to a given location
  * @param string $location location to be redirected to
  */
-function redirect($location){
+function redirect($location)
+{
     header(sprintf('Location: %s', $location));
     die();
 }
@@ -404,11 +455,217 @@ function redirect($location){
  * Get current user id
  * @return bool current user id or False if not logged in
  */
-function get_user_id(){
+function get_user_id()
+{
     session_start();
-    if (isset($_SESSION['user_id'])){
+    if (isset($_SESSION['user_id'])) {
         return $_SESSION['user_id'];
     } else {
         return False;
+    }
+}
+
+/**
+ * Generates an array with the first- and last name
+ * that belong to a given user_id
+ * @param object $pdo db object
+ * @param int $user_id id from the serie
+ * @return mixed
+ */
+function get_user_names($pdo, $user_id)
+{
+    $stmt = $pdo->prepare('SELECT firstname , lastname FROM users WHERE id = ?');
+    $stmt->execute([$user_id]);
+    $user_info = $stmt->fetch();
+    $user_info_exp = Array();
+
+    /* Create array with htmlspecialchars */
+    foreach ($user_info as $key => $value) {
+        $user_info_exp[$key] = htmlspecialchars($value);
+    }
+    return $user_info_exp;
+}
+
+/**
+ * Registers the user and logs the user in when successful
+ * @param object $pdo db object
+ * @param array $form_data with POST data
+ * @return array
+ */
+function register_user($pdo, $form_data)
+{
+    /* Check if all fields are set */
+    if (
+        empty(
+        $form_data['username']) or
+        empty($form_data['password']) or
+        empty($form_data['firstname']) or
+        empty($form_data['lastname'])
+    ) {
+        return [
+            'type' => 'danger',
+            'message' => 'You should enter a username, password, first- and last name.'
+        ];
+    }
+
+    /* Check if user already exists */
+    try {
+        $stmt = $pdo->prepare('SELECT * FROM users WHERE username = ?');
+        $stmt->execute([$form_data['username']]);
+        $user_exists = $stmt->rowCount();
+    } catch (\PDOException$e) {
+        return ['type' => 'danger',
+            'message' => sprintf('There was an error: %s', $e->getMessage())
+        ];
+    }
+
+    /* Return error message for existing username */
+    if (!empty($user_exists)) {
+        return [
+            'type' => 'danger',
+            'message' => 'The username you entered does already exists!'
+        ];
+    }
+
+    /* Hash password */
+    $password = password_hash($form_data['password'], PASSWORD_DEFAULT);
+
+    /* Save user to the database */
+    try {
+        $stmt = $pdo->prepare('INSERT INTO users (username, password, firstname,lastname) VALUES (?, ?, ?, ?)');
+        $stmt->execute([$form_data['username'], $password, $form_data['firstname'], $form_data['lastname']]);
+        $user_id = $pdo->lastInsertId();
+    } catch (PDOException$e) {
+        return [
+            'type' => 'danger',
+            'message' => sprintf('There was an error: %s', $e->getMessage())
+        ];
+    }
+
+    /* Login user and redirect */
+    session_start();
+    $_SESSION['user_id'] = $user_id;
+    $feedback = [
+        'type' => 'success',
+        'message' => sprintf("%s your account was successfully created!",
+            get_username($pdo, $_SESSION['user_id']))
+    ];
+    redirect(sprintf('/DDWT18/week2/myaccount/?error_msg=%s',
+        json_encode($feedback)));
+}
+
+/**
+ * Generates an array with the username
+ * that belong to a given user_id
+ * @param object $pdo db object
+ * @param int $user_id
+ * @return mixed
+ */
+function get_username($pdo, $user_id)
+{
+    $stmt = $pdo->prepare('SELECT username FROM users WHERE id = ?');
+    $stmt->execute([$user_id]);
+    $username = $stmt->fetch();
+    $username_exp = Array();
+
+    /* Create array with htmlspecialchars */
+    foreach ($username as $key => $value) {
+        $username_exp[$key] = htmlspecialchars($value);
+    }
+    return $username_exp;
+}
+
+/**
+ * Log a user in
+ * @param object $pdo db object
+ * @param mixed $form_data POST data with user info
+ * @return mixed
+ */
+function login_user($pdo, $form_data)
+{
+    /* Check if all fields are set */
+    if (
+        empty($form_data['username']) or
+        empty($form_data['password'])
+    ) {
+        return [
+            'type' => 'danger',
+            'message' => 'You should enter a username and password.'
+        ];
+    }
+
+    /* Check if user exists */
+    try {
+        $stmt = $pdo->prepare('SELECT * FROM users WHERE username = ?');
+        $stmt->execute([$form_data['username']]);
+        $user_info = $stmt->fetch();
+    } catch (\PDOException$e) {
+        return [
+            'type' => 'danger',
+            'message' => sprintf('There was an error: %s', $e->getMessage())
+        ];
+    }
+
+    /* Return error message for wrong username */
+    if (empty($user_info)) {
+        return [
+            'type' => 'danger',
+            'message' => 'The username you entered does not exist!'
+        ];
+    }
+
+    /* Check password */
+    if (!password_verify($form_data['password'], $user_info['password'])) {
+        return [
+            'type' => 'danger',
+            'message' => 'The password you entered is incorrect!'
+        ];
+    } else {
+        session_start();
+        $_SESSION['user_id'] = $user_info['id'];
+        $feedback = [
+            'type' => 'success',
+            'message' => sprintf('%s, you were logged in successfully!',
+                get_username($pdo, $_SESSION['user_id']))
+        ];
+        redirect(sprintf('/DDWT18/week2/myaccount/?error_msg=%s',
+            json_encode($feedback)));
+        $_SESSION['user_id'] = $user_info['id'];
+    }
+}
+
+/**
+ * Checks if the user is logged in
+ * @return bool True if logged in
+ */
+function check_login()
+{
+    session_start();
+    if (isset($_SESSION['user_id'])) {
+        return True;
+    } else {
+        return False;
+    }
+}
+
+/**
+ * Logs the user out
+ * @return array message if logged out successfully
+ */
+function logout_user()
+{
+    try {
+        session_start();
+        $_SESSION = array();
+        session_destroy();
+        return [
+            'type' => 'success',
+            'message' => 'You logged out successfully.'
+        ];
+    } catch (Exception $e) {
+        return [
+            'type' => 'danger',
+            'message' => sprintf('Logging out went wrong. Error: ', $e->getMessage())
+        ];
     }
 }
